refactoring, docs
