+++ designer/server/src/test/scala/pl/touk/nussknacker/ui/api/BaseDeploymentApiHttpServiceBusinessSpec.scala
@@ -153,4 +153,14 @@ trait BaseDeploymentApiHttpServiceBusinessSpec extends WithFlinkContainersDeploy
     }
   }
 
+  protected def listJobManagerFiles(containerPath: String): List[String] = {
+    val result = jobManagerContainer.container.execInContainer("ls", "-Atr1p", containerPath)
+    if (result.getExitCode != 0) {
+      throw new IllegalStateException(
+        s"Failed to list $containerPath in JobManager container: ${result.getStderr}"
+      )
+    }
+    result.getStdout.split("\n").iterator.map(_.trim).filter(_.nonEmpty).toList
+  }
+
 }
